+++ code/system.ui/react.common/src/common/types.mts
@@ -5,3 +5,4 @@ export * from '../types.mjs';
  */
 export type { DomRect } from 'sys.types';
 export type { CssValue } from 'sys.util.css';
+export type { DevCtx } from 'sys.ui.react.dev.types';
