+++ code/system.ui/react.dev.types/src/index.mts
@@ -1,1 +1,1 @@
-export * from './types.mjs';
+export * from './types';
