--- code/system.ui/react.dev.types/src/test/index.ts
@@ -1,3 +1,0 @@
-export { expect, expectError } from 'sys.test';
-export { describe, it } from 'vitest';
-export * from '../common';
