--- code/system.ui/react.dev.types/src/types.mts
@@ -1,4 +1,0 @@
-export {};
-
-// TEMP 🐷
-export type DevCtx = { TODO: number };
