+++ code/system.ui/react.dev.types/src/types/common.ts
@@ -1,0 +1,2 @@
+import * as t from './common.types.mjs';
+export { t };
