+++ code/system.ui/react.dev.types/src/types/index.ts
@@ -1,0 +1,1 @@
+export * from './types.mjs';
