+++ code/system.ui/react.dev.types/src/types/types.mts
@@ -1,0 +1,6 @@
+export * from './types.mjs';
+export * from './types.bus.mjs';
+export * from './types.ctx.mjs';
+
+export type MarginInput = number | [number] | [number, number] | [number, number, number, number];
+export type Margin = [number, number, number, number];
