--- code/system.ui/react.dev/src/Spec/types.mts
@@ -1,1 +1,0 @@
-export {};
