+++ code/system.ui/react.dev/src/types.mts
@@ -1,8 +1,3 @@
-export * from './Spec/types.mjs';
-
-export * from './logic.Bus/types.mjs';
-export * from './logic.Ctx/types.mjs';
-
 export * from './ui/Entry/types.mjs';
 export * from './ui/Entry.SpecList/types.mjs';
