+++ code/system/sys.types/src/types/types.mts
@@ -1,0 +1,1 @@
+export type IgnoredResponse = any | Promise<any>;
